- Analyzed Percunia's entire code base and fixed a large number of memory leaks and related problems. - Added a second Pecunia executable project entry which enables extended memory tracking (malloc stack, zombies). - Removed a few unused code lines + added some initial values to vars that could be used without initialization. - StatementsListview: hide top labels if no header is shown to avoid ghost values.

+++ client/Source/CategoryRepWindowController.m
@@ -617,12 +617,12 @@ static NSString* const PecuniaHitNotification = @"PecuniaMouseHit";
     switch ([(NSNumber*)pieChart.identifier intValue])
     {
         case EARNINGS_PLOT_ID:
-            if (index == earningsExplosionIndex) {
+            if ((NSInteger)index == earningsExplosionIndex) {
                 result = 10.0;
             }
             break;
         case SPENDINGS_PLOT_ID:
-            if (index == spendingsExplosionIndex) {
+            if ((NSInteger)index == spendingsExplosionIndex) {
                 result = 10.0;
             }
             break;
